backend/requirements.txt
python-nmap
scapy
pymetasploit3